changes in preparation for scala-native support
