+++ project/build.properties
@@ -1,1 +1,2 @@
-sbt.version=1.9.9
+#sbt.version=1.9.9
+sbt.version=1.10.0
