+++ project/plugins.sbt
@@ -2,10 +2,10 @@ scalaVersion := "2.12.19"
 
 val SONATYPE_VERSION = sys.env.getOrElse("SONATYPE_VERSION", "3.10.0") // "3.9.21")
 
-//addSbtPlugin("org.scala-native" % "sbt-scala-native" % "0.4.17")
+//addSbtPlugin("org.scala-native" % "sbt-scala-native" % "0.5.1")
 
 addSbtPlugin("ch.epfl.scala"  % "sbt-bloop"     % "1.5.17")
-addSbtPlugin("ch.epfl.scala"  % "sbt-scalafix"  % "0.12.0")
+addSbtPlugin("ch.epfl.scala"  % "sbt-scalafix"  % "0.12.1")
 addSbtPlugin("org.xerial.sbt" % "sbt-sonatype"  % SONATYPE_VERSION)
 addSbtPlugin("org.scalameta"  % "sbt-scalafmt"  % "2.5.2")
 addSbtPlugin("com.github.sbt" % "sbt-dynver"    % "5.0.1")
